requirements.txt
# Core Application
pydantic
pydantic-settings
python-dotenv
jsonschema

# AI & Web Interaction
google-generativeai
playwright
beautifulsoup4
jinja2
Pillow